backend/requirements.txt
fastapi[standard]==0.113.0
pydantic==2.8.0
numpy==2.2.0
opencv-contrib-python==4.11.0.86
pillow==11.2.1